Reworked comprehension handling
* Streamlined how comprehensions are handled throughout both the
`api_fuzzer` and `set_meta_tests` specs. Comprehensions now have a
general structure of `<type=gen_method=description>`, where `<`, `=` and
`>` are delimiters subject to change, `type` is related to the resulting
type of the comprehension, `gen_method` represents an abstract way of
obtaining the eventual object, and `description` provides some
additional context about retrieving the object.
* Also streamlined internal representation of the comprehensions;
currently using the `ExplicitType` class, but might move away from it in
the future