Fuzzer bug fix and additional documentation
* Rather than declaring and initializing objects
whenever one was created, we now create instances in the background and
only initialize them on demand, when they are used. This allows for
finer controlled generation paradigms of ApiObjects.
* Fixed a bug with fuzzer generation when moved to an "on-demand" object
declaration logic. Objects were not initialized in certain edge cases.
* Moved the declaration determination logic inside ApiInstructions,
such that whenever an ApiInstruction which initializes an object is
emitted, the object the is stored to is set as declared.
* Added more documentation throughout